Move PathItem to Objects folder
